Buggy phasers routine added!

+++ src/mymath.h
@@ -2,7 +2,7 @@
 #ifndef mymath_h
 #define mymath_h
 
-long r_GetDistance(short xdiff, short ydiff, short zdiff);
+long r_GetDistance(int xdiff, int ydiff, int zdiff);
 signed char byteCos(unsigned char x);
 signed char byteSin(unsigned char x);
